# api/openapi.yaml
openapi: 3.0.3
info:
  title: Hotel Reviews Microservice API
  description: |
    A comprehensive microservice for managing hotel reviews, ratings, and related data.
    
    ## Features
    - Hotel and provider management
    - Review creation, retrieval, and management
    - Advanced search and filtering capabilities
    - File upload and batch processing
    - User authentication and authorization
    - Real-time metrics and monitoring
    
    ## Authentication
    This API uses JWT Bearer token authentication. Obtain a token by calling the `/auth/login` endpoint.
    
    ## Rate Limiting
    API requests are rate-limited. Check the `X-RateLimit-*` headers in responses for current limits.
    
    ## Versioning
    The API uses URL path versioning (e.g., `/api/v1/`). Current version is v1.
  version: 1.0.0
  contact:
    name: Hotel Reviews API Support
    email: support@hotelreviews.com
    url: https://docs.hotelreviews.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://hotelreviews.com/terms
servers:
  - url: https://api.hotelreviews.com/api/v1
    description: Production server
  - url: https://staging-api.hotelreviews.com/api/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
    description: Development server
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email verification
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
            examples:
              user_registration:
                summary: Standard user registration
                value:
                  username: "johndoe"
                  email: "john.doe@example.com"
                  password: "SecurePassword123!"
                  first_name: "John"
                  last_name: "Doe"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                registration_success:
                  summary: Successful registration
                  value:
                    success: true
                    data:
                      user:
                        id: "123e4567-e89b-12d3-a456-426614174000"
                        username: "johndoe"
                        email: "john.doe@example.com"
                        first_name: "John"
                        last_name: "Doe"
                        is_verified: false
                      access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      refresh_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      expires_in: 3600
                    timestamp: "2024-01-15T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and obtain access tokens
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              login_with_email:
                summary: Login with email
                value:
                  email: "john.doe@example.com"
                  password: "SecurePassword123!"
              login_with_username:
                summary: Login with username
                value:
                  username: "johndoe"
                  password: "SecurePassword123!"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Obtain a new access token using a refresh token
      operationId: refreshToken
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  description: Valid refresh token
              required:
                - refresh_token
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate current access and refresh tokens
      operationId: logoutUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Review Endpoints
  /reviews:
    get:
      tags:
        - Reviews
      summary: List reviews
      description: Retrieve a paginated list of reviews with optional filtering
      operationId: listReviews
      security:
        - BearerAuth: []
      parameters:
        - name: hotel_id
          in: query
          description: Filter reviews by hotel ID
          schema:
            type: string
            format: uuid
        - name: provider_id
          in: query
          description: Filter reviews by provider ID
          schema:
            type: string
            format: uuid
        - name: rating_min
          in: query
          description: Minimum rating filter
          schema:
            type: number
            minimum: 1
            maximum: 5
        - name: rating_max
          in: query
          description: Maximum rating filter
          schema:
            type: number
            minimum: 1
            maximum: 5
        - name: language
          in: query
          description: Filter by review language
          schema:
            type: string
            pattern: '^[a-z]{2}$'
            example: "en"
        - name: from_date
          in: query
          description: Filter reviews from this date
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          description: Filter reviews to this date
          schema:
            type: string
            format: date
        - name: limit
          in: query
          description: Number of reviews to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of reviews to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
            enum: [rating, review_date, created_at, helpful_votes]
            default: created_at
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListResponse'
              examples:
                review_list:
                  summary: List of reviews
                  value:
                    success: true
                    data:
                      reviews:
                        - id: "550e8400-e29b-41d4-a716-446655440000"
                          hotel_id: "123e4567-e89b-12d3-a456-426614174000"
                          provider_id: "987fcdeb-51a2-4567-8901-23456789abcd"
                          rating: 4.5
                          title: "Great stay!"
                          comment: "Had a wonderful time at this hotel."
                          review_date: "2024-01-10T00:00:00Z"
                          language: "en"
                          is_verified: true
                          helpful_votes: 12
                          total_votes: 15
                          created_at: "2024-01-11T10:30:00Z"
                      pagination:
                        total: 156
                        limit: 20
                        offset: 0
                        has_next: true
                        has_previous: false
                    timestamp: "2024-01-15T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Reviews
      summary: Create a new review
      description: Create a new hotel review
      operationId: createReview
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
            examples:
              basic_review:
                summary: Basic review
                value:
                  hotel_id: "123e4567-e89b-12d3-a456-426614174000"
                  reviewer_info_id: "456e7890-e89b-12d3-a456-426614174001"
                  rating: 4.5
                  title: "Great stay!"
                  comment: "Had a wonderful time at this hotel. Staff was friendly and rooms were clean."
                  review_date: "2024-01-10T00:00:00Z"
                  language: "en"
              detailed_review:
                summary: Detailed review with all ratings
                value:
                  hotel_id: "123e4567-e89b-12d3-a456-426614174000"
                  reviewer_info_id: "456e7890-e89b-12d3-a456-426614174001"
                  rating: 4.2
                  title: "Good hotel with excellent service"
                  comment: "Overall a good experience. The location was perfect and staff very helpful."
                  review_date: "2024-01-10T00:00:00Z"
                  language: "en"
                  service_rating: 5.0
                  cleanliness_rating: 4.0
                  location_rating: 5.0
                  value_rating: 3.5
                  comfort_rating: 4.0
                  facilities_rating: 4.0
                  trip_type: "business"
                  room_type: "standard"
      responses:
        '201':
          description: Review created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /reviews/{id}:
    get:
      tags:
        - Reviews
      summary: Get review by ID
      description: Retrieve a specific review by its ID
      operationId: getReview
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Review ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Review retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Reviews
      summary: Update review
      description: Update an existing review (only by owner or admin)
      operationId: updateReview
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Review ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReviewRequest'
      responses:
        '200':
          description: Review updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags:
        - Reviews
      summary: Delete review
      description: Delete a review (only by owner or admin)
      operationId: deleteReview
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Review ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Review deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /reviews/upload:
    post:
      tags:
        - Reviews
      summary: Upload reviews file
      description: Upload a CSV file containing reviews for batch processing
      operationId: uploadReviewsFile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file containing reviews
                provider_id:
                  type: string
                  format: uuid
                  description: Provider ID for the reviews
                validate_only:
                  type: boolean
                  default: false
                  description: Only validate the file without processing
              required:
                - file
                - provider_id
      responses:
        '202':
          description: File uploaded and processing started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /hotels:
    get:
      tags:
        - Hotels
      summary: List hotels
      description: Retrieve a paginated list of hotels with optional filtering
      operationId: listHotels
      security:
        - BearerAuth: []
      parameters:
        - name: name
          in: query
          description: Filter by hotel name (partial match)
          schema:
            type: string
            maxLength: 255
        - name: city
          in: query
          description: Filter by city
          schema:
            type: string
            maxLength: 100
        - name: country
          in: query
          description: Filter by country
          schema:
            type: string
            maxLength: 100
        - name: star_rating
          in: query
          description: Filter by star rating
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: min_rating
          in: query
          description: Minimum average review rating
          schema:
            type: number
            minimum: 1
            maximum: 5
        - name: limit
          in: query
          description: Number of hotels to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of hotels to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Hotels retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Hotels
      summary: Create a new hotel
      description: Create a new hotel (admin only)
      operationId: createHotel
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHotelRequest'
      responses:
        '201':
          description: Hotel created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /providers/{id}:
    get:
      tags:
        - Providers
      summary: Get provider by ID
      description: Retrieve a specific provider by its ID
      operationId: getProvider
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Provider ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Provider retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Health and Monitoring Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from the authentication endpoints
  schemas:
...
    UserRegistrationRequest:
      type: object
      required:
        - username
        - email
        - password
        - first_name
        - last_name
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          pattern: '^[a-zA-Z0-9_]+$'
          description: Unique username
        email:
          type: string
          format: email
          maxLength: 255
          description: User email address
        password:
          type: string
          minLength: 8
          maxLength: 128
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]'
          description: Password with at least 8 characters, including uppercase, lowercase, number, and special character
        first_name:
          type: string
          maxLength: 100
          description: User's first name
        last_name:
          type: string
          maxLength: 100
          description: User's last name
...